+++ CHANGELOG.md
@@ -1,11 +1,20 @@
-# 0.1.0 (2025-11-20)
+# 0.4.0 (2025-12-02)
+
+
+### Bug Fixes
+
+* copy ocr model to correct path ([34d9e1f](https://github.com/runslikebutter/license-plate-recorder/commit/34d9e1f5e0ac3ac3b329f057fc88a9861b1b449e))
+* include car_lpd_11nv6.engine in Docker build context ([c7826ee](https://github.com/runslikebutter/license-plate-recorder/commit/c7826ee89e5003762ead8ad17cefed8fa3a36778))
+* update workflows to checkout correct branch ([aa3f4ec](https://github.com/runslikebutter/license-plate-recorder/commit/aa3f4ec94f7a472305c2d6835af292de1098c837))
 
 
 ### Features
 
 * add ocr prod models and configs ([81a6b50](https://github.com/runslikebutter/license-plate-recorder/commit/81a6b50a73a2dbc06aab2d5bd4e2a536eb5a1788))
+* add reconnection logic ([5204a6a](https://github.com/runslikebutter/license-plate-recorder/commit/5204a6a5727e9e67f46ec4297cca7cf5fba5f6c0))
 * license recorder add implementation ([42d198d](https://github.com/runslikebutter/license-plate-recorder/commit/42d198ddd362ecc542fbed486ffe0639a7c58531))
 * refactor to use trt models ([be9fb80](https://github.com/runslikebutter/license-plate-recorder/commit/be9fb805cfd24fa006d3c5846239e50c8a0d94c5))
+* upload recordings to s3 ([11fd1dd](https://github.com/runslikebutter/license-plate-recorder/commit/11fd1ddb335d9a674007e8ff8f43937de3145315))